Merged the basic lib into the extended-basis lib.
The merge was done because it became apparent (to me) that it is just too
painful to have them separated.  One problem is that there were many
little utilities in the basic lib that I wanted to use in the
extended-basis lib while implementing new extensions (will appear later)
and I don't want to keep duplicating stuff.

This (at least temporarily) breaks the build files on compilers except
MLton (and possibly MLKit).  It seems that it will take some work to
support SML/NJ's CM, but I definitely plan to do it as soon as I figure
out how to do it least painfully.


SVN r4845