The beginnings of a new approach to Async based on "deferred" values and streams.  A deferred value is a value that will be determined at some point (i.e. asynchronously).  It is different from a CML event in two ways:
  1. It is not a potential communication that must be committed to in
     order to be enabled.  Rather, it represents a computation that
     has already been committed to (i.e. begun), and may yield a value
     at some point.
  2. Once it becomes determined with a value, it always has that
     value. 

The deferred type more accurately reflects the result of Ivar.read,
since an ivar never changes once it is written.  It will make code
clearer to use this more precise type to capture write-once events.

A "stream" is the natural asynchronous analogue of a list -- the only
difference is that when getting the first element, one cannot be sure
whether it has been computed or not, so one receives a deferred value
instead of a normal one.  Streams naturally represent the result of
reading from a multicast -- each reader gets his own stream (i.e.
a read-only copy of the rest of the values).  They also make it
trivial to implement multicast -- all readers share some tail of the
same stream.

The idea going forward is that events, i.e. potential communications
that vary in time, represent potential deferred values, and that one
"commits" to an event using

  val commit: 'a Event.t -> 'a Deferred.t

The notion of commiting was implicit in CML and with the old Async,
but it was tied with actually getting the value.  I think it will be
clearer to make the notion explicit.



SVN r5444