Renamed the plus val of MONADP_CORE to <|> and introduced a fixity declaration for it (just below of the sequencing monadic operators).
The rationale for choosing the symbolic identifier <|>.  Currently all
other monadic operators use >> as a prefix.  The < and > symbols in <|>
make it visually belong with the other monadic operators and reflect the
semantics (evaluation can go either way) of the operator.  The symbol | is
not used in any operator of the Basis or previously in any operator of the
Extended Basis.  This means that there is no previous precedence
associated with it.  (Operators that are variations of a "base" operator
like + or * are generally given the same precedence in the Extended
Basis.)  In SML, the symbol | is used with pattern matching where it is
used to separate alternative clauses.  This also reminds about the
semantics of plus in MonadPlus.  Finally, the monadic parser combinator
library Parsec uses <|>.

Also simplified the implementation of seq/app functions.


SVN r5449