Moved isomorphism lifters from Iso to other modules and added a few more lifters.  Isomorphisms are quite hand in generic programming.
Also changed the type of the Fn.iso lifter (compared to the now removed
Iso.-->).  I'm not sure whether it is ultimately better to reflect the
contravariance in the type or not, but it seems that handling the
contravariance once in Fn.iso is better than swapping isomorphisms
everywhere when Fn.iso is used.


SVN r5806