Delay HashMap update after rd/wr when dealing with acyclic references.
As noted by Elsman [http://mlton.org/References#Elsman04], Standard ML
does not provide a means of extracting a hashable identity from a mutable
object.  Regardless of identity, structurally equivalent mutable objects
hash to the same value.  This increases the asymptotic complexity of
algorithms, including pickling, that need to distinguish mutable objects
by identity.  Elsman then described two optimized combinators for pickling
references whose use presents additional proof obligations to programmers.
One of the optimized combinators, ref0, simply assumes that no cycles
appear through references pickled with it.  Perhaps surprisingly, it is
possible to perform a similar optimization automatically.  Specifically,
during type representation construction, it is possible to compute a
conservative approximation as to whether values of a mutable type may be a
part of a cycle or not.  An algorithm for this boils down to tracking down
whether a value of the given mutable type may contain values of
(exceptions or) recursive datatypes that may contain a value of the
mutable type.  Our library includes the DataRecInfo generic, which
implements such an algorithm.


SVN r5872