Committed some experimental changes that have been pending for a while.
Added whenSeq and whenArb for committing to a list of events.

Added a binary choice combinator <|> and a never (zero) event and
reimplemented choose using them.

Changed the tests to use the new generic and unit-test libraries instead
of the mist-util library, which is on its way out.


SVN r5948