More precise (but asymptotically slower) hashing of exceptions.  Some minor tweaks.  Fixed to properly decrease maxDepth only at (datatype and exception) constructors and only when maxDepth > 0.
BTW, it would be possible to perform exception lookup asymptotically
faster (roughly O(1) vs O(n), where n is the number of registered
exception constructors) via General.exnName, but it is going to have to
wait.  The assumption is that use of generics to process exceptions is
very rare and there are more important issues to fix.


SVN r5971