Avoid constructing a proxy ref cell (holding a dummy value) when unpickling acyclic refs.
A similar optimization could also be done for arrays, but it is probably
more efficient to construct an array holding dummy values and then update
the array incrementally rather than accumulate a list of values and then
construct an array from the list.

This could also be seen as a correctness / semantics issue.  The
construction of a dummy value (using the Some generic) might actually
fail.  So, it makes sense to avoid constructing dummy values when they are
not strictly needed.


SVN r6017