Redesigned the unit test framework and random testing interface in particular.
The old random testing interface was largely a copy of QuickCheck.  The
main difference is that there is no separate concept of a "property" or
"law".  A test, whether randomized or not, is an effectful procedure that
raises an exception upon failure.  In general, this reduces or eliminates
differences between randomized and non-randomized tests.  A particular
benefit of this approach is that the same assertion procedures can be used
in both randomized and non-randomized tests --- there is no need to
provide two sets of assertion procedures.

Another notable difference compared to the old design is that data
collection, for debugging randomized tests, is no longer an integral part
of the framework.  In the original QuickCheck design, data collection is
an integral part of the framework, because all data must be passed through
the property combinators.  In SML, we can instead use side-effects to
avoid having to bundle data collection with the framework.  This
simplifies the implementation considerably.

The main disadvantage of the new design seems to be a small increase in
verbosity (an extra thunk in some cases and more work to collect
statistics).

There is also no longer need to reveal that random generators are
functions.


SVN r6051