Switched to using a slightly simpler monad ("Pass" instead of State) for passing the pickling environment.
Introduced a datatype for the monad transformer to workaround a bug in
MLKit (rev 2287).  This also significantly reduced the amount of code
generated by SML/NJ.

Switched from pattern matching to equality comparison for IntInf values as
a workaround for a bug in MLKit (rev 2287).

Some minor tweaks.


SVN r6103