Eliminated the use of the DataRecInfo generic from the implementation of the pickling generic.  DataRecInfo performs a simple data recursion analysis, which was used in the pickling generic to automatically perform an optimization similar to the {ref0} combinator described in
   Type-specialized serialization with sharing.
   Martin Elsman.
   In Sixth Symposium on Trends in Functional Programming (TFP'05),
   September 2005.

Unfortunately, the data recursion analysis is expensive and cannot
typically be constant folded by compilers.  Furthermore, the optimization
allowed by the analysis is rather insignificant.  It allows to avoid
creating a dummy value and to delay the insertion of the ref or array into
the pickling environment.  Leaving the analysis out and just treating all
refs and arrays as potentially cyclic seems to be a better approach.


SVN r6342