Fixed not to refer to a non-existent (removed) file.
Interestingly, I did not notice this problem earlier, because SML/NJ did
not give an error for it.  I assume SML/NJ's CM, having compiled the file
earlier, just assumed it was still there.  Later, when something else
changed, the file would need to be recompiled and SML/NJ noticed that the
file was no longer there.  I would call this a bug in CM.  I have not
tried to reproduce this.


SVN r6348