Enhanced read to also parse tuples written using the record syntax.
Replaced the use of a list of functions with an incrementally built
function in the type representations of products in the read generic.
This seems to dramatically improve the generated code with MLton.  The
reasons for this seem straightforward.  Many crucial optimizations
(inlining, contification, ...) do not work across a list of functions.

Made some other minor performance tweaks in the Read generic.

Also made the Record and Tuple types in the GENERICS signature concrete
and mapped them to String and Int, respectively.  This is used in the Read
generic for a minor performance benefit.  While this is somewhat of an
ugly hack, note that outside of defining a generic function, the types
cannot be used for anything.  What matters is that they are different
types.


SVN r6375