Changed generic hash to produce a Word32 hash value rather than Word.
This has the benefit that the result will be the same on all compilers and
that can be valuable for a number of purposes.  The disadvantage is that
one often wants the hash value to be a Word, but it is not a big deal to
convert with Word32.toWord.

Another alternative would have been to implement a two generic hash
functions (one for producing platform independent hashes and another for
native hashes), but it does not seem worth it.


SVN r6926